Accommodate 303 in /me/following/:id
The API endpoint /me/following/:id returns 404 (Not Found) if
authenticated user does not follow :id. If authenticated user follows
:id, it returns 303 (See Other), redirecting to /users/:id. Because the
response is not 2XX (Success), the client returns ResponseError.

This change allows 2XX (Success) and 3XX (Redirection) - not returning
ResponseError.

When redirection occurs, it is my understanding that authentication
information is not passed on to the new Location - resulting in 401 (Not
Authorized). So this change also sets follow_redirects to false, so
that redirection does not happen